refactor(cli): adapt scan runner to pipeline
